Add files via upload
Every month the U.S. Bureau of Labor Statistics publishes employment figures for large sections of the U.S.
Economy. The task is to forecast the amount of people employed in the financial activities sector for four
years. This information is publicly reported by the St. Louis FRED and can be found at the following link.
https://fred.stlouisfed.org/series/USFIRE.
Four years represents a large amount of time and confidence windows will be wide. Therefore, I will test a 
few forecasting models by comparing their respective RMSE and Winkler
scores on a known portion of the time series.